requirements.txt
airtable-python-wrapper
